// projects/utils/ui-kit/definitions/components/banner/banner.type.ts
export enum UkBannerTopStatus {
  OFF = 'OFF',
  H_60 = 'H-60',
  H_150 = 'H-150',
}
export type BannerTopStatus = `${UkBannerTopStatus}`;

export enum UkBannerContentStartStatus {
  OFF = 'OFF',
  H_50_SQUARE = 'H-50-SQUARE',
  H_50_CIRCLE = 'H-50-CIRCLE',
  H_65_SQUARE = 'H-65-SQUARE',
}
export type UBannerContentStartStatus = `${UkBannerContentStartStatus}`;