// projects/utils/ui-kit/definitions/components/divider/divider.type.ts
export enum UkDividerBgColor {
  GRADE_1 = 'GRADE-1',
  GRADE_2 = 'GRADE-2',
  HIGH_EMPHASIS = 'HIGH-EMPHASIS',
}
export type DividerBgColor = `${UkDividerBgColor}`;

export enum UkDividerSize {
  PX_1 = 'PX-1',
  PX_2 = 'PX-2',
}
export type DividerSize = `${UkDividerSize}`;

export enum UkDividerMargin {
  NONE = 'NONE',
  M_4 = 'M-4',
  M_8 = 'M-8',
  M_16 = 'M-16',
}
export type DividerMargin = `${UkDividerMargin}`;

export enum UkDividerDisplay {
  HORIZONTAL = 'HORIZONTAL',
  VERTICAL = 'VERTICAL',
}
export type DividerDisplay = `${UkDividerDisplay}`;