// projects/utils/ui-kit/definitions/components/page/page.type.ts
export enum UkPageBackgroundColor {
  WHITE = 'WHITE',
  TRANSPARENT = 'TRANSPARENT',
  SECONDARY_900 = 'SECONDARY-900',
  SECONDARY_800 = 'SECONDARY-800',
  SECONDARY_TRANSPARENT_50 = 'SECONDARY-TRANSPARENT-50',
}
export type PageBackgroundColor = `${UkPageBackgroundColor}`;

export enum UkPageBodyOverFlow {
  UNSET = 'UNSET',
  SCROLL = 'SCROLL',
  HIDDEN = 'HIDDEN',
}
export type PageBodyOverFlow = `${UkPageBodyOverFlow}`;