// projects/utils/ui-kit/definitions/components/select/select.type.ts
export enum UkSelectBgColor {
  NONE = 'NONE',
  BACKGROUND_SURFACE_00 = 'BACKGROUND-SURFACE-00',
  BACKGROUND_PRIMARY = 'BACKGROUND-PRIMARY',
  BACKGROUND_SECONDARY = 'BACKGROUND-SECONDARY',
}
export type SelectBgColor = `${UkSelectBgColor}`;

export enum UkSelectBorderColor {
  TRANSPARENT = 'TRANSPARENT',
  GRADE_1 = 'GRADE-1',
  GRADE_2 = 'GRADE-2',
  REFERENCE_PRIMARY_MAIN = 'REFERENCE-PRIMARY-MAIN',
  CONTENT_ERROR = 'CONTENT-ERROR',
}
export type SelectBorderColor = `${UkSelectBorderColor}`;

export enum UkSelectBorderRadius {
  NONE = 'NONE',
  LARGER = 'LARGER',
  LARGE = 'LARGE',
  MEDIUM = 'MEDIUM',
  SMALL = 'SMALL',
  SMALLER = 'SMALLER',
}
export type SelectBorderRadius = `${UkSelectBorderRadius}`;

export enum UkSelectFgColor {
  FG_MILK_800 = 'FG-MILK-800',
  FG_SMOKE_700 = 'FG-SMOKE-700',
}
export type SelectFgColor = `${UkSelectFgColor}`;

export enum UkSelectMarginRight {
  NONE = 'NONE',
  M10 = 'M10',
}
export type SelectMarginRight = `${UkSelectMarginRight}`;